--- src/main/java/org/nanotek/crawler/data/config/meta/RelationType.java
@@ -1,6 +1,0 @@
-package org.nanotek.crawler.data.config.meta;
-
-public enum RelationType{
-	ONE,
-	MANY
-}
