+++ src/main/java/org/nanotek/crawler/data/domain/meta/IClass.java
@@ -1,4 +1,4 @@
-package org.nanotek.crawler.data.config.meta;
+package org.nanotek.crawler.data.domain.meta;
 
 import java.util.List;
