+++ src/main/java/org/nanotek/crawler/data/domain/meta/Id.java
@@ -1,4 +1,4 @@
-package org.nanotek.crawler.data.config.meta;
+package org.nanotek.crawler.data.domain.meta;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
