+++ src/main/java/org/nanotek/crawler/data/domain/meta/MetaIdentity.java
@@ -1,4 +1,4 @@
-package org.nanotek.crawler.data.config.meta;
+package org.nanotek.crawler.data.domain.meta;
 
 import java.util.ArrayList;
 import java.util.List;
