+++ src/main/java/org/nanotek/crawler/data/domain/meta/MetaRelationClass.java
@@ -1,4 +1,4 @@
-package org.nanotek.crawler.data.config.meta;
+package org.nanotek.crawler.data.domain.meta;
 
 import java.util.Iterator;
 import java.util.List;
