+++ src/main/java/org/nanotek/crawler/data/domain/meta/RelationType.java
@@ -1,0 +1,6 @@
+package org.nanotek.crawler.data.domain.meta;
+
+public enum RelationType{
+	ONE,
+	MANY
+}
