+++ src/main/java/org/nanotek/crawler/data/domain/meta/TempClass.java
@@ -1,4 +1,4 @@
-package org.nanotek.crawler.data.config.meta;
+package org.nanotek.crawler.data.domain.meta;
 
 public class TempClass {
 	String source;
