+++ src/main/java/org/nanotek/crawler/data/util/db/support/PostProcessor.java
@@ -1,6 +1,6 @@
 package org.nanotek.crawler.data.util.db.support;
 
-import org.nanotek.crawler.data.config.meta.IClass;
+import org.nanotek.crawler.data.domain.meta.IClass;
 
 public interface PostProcessor<T> {
