+++ src/main/java/org/nanotek/crawler/service/GraphRelationsConfig.java
@@ -22,9 +22,9 @@ import org.jgrapht.alg.shortestpath.YenKShortestPath;
 import org.jgrapht.graph.builder.GraphTypeBuilder;
 import org.nanotek.crawler.Base;
 import org.nanotek.crawler.data.SearchParameters;
-import org.nanotek.crawler.data.config.meta.MetaClass;
-import org.nanotek.crawler.data.config.meta.MetaEdge;
-import org.nanotek.crawler.data.config.meta.TempClass;
+import org.nanotek.crawler.data.domain.meta.MetaClass;
+import org.nanotek.crawler.data.domain.meta.MetaEdge;
+import org.nanotek.crawler.data.domain.meta.TempClass;
 import org.nanotek.crawler.data.domain.mongodb.ClassPath;
 import org.nanotek.crawler.data.mongo.repositories.ClassPathRepository;
 import org.nanotek.crawler.data.stereotype.EntityBaseRepository;
