+++ src/main/java/org/nanotek/crawler/service/SimplePathValidator.java
@@ -2,7 +2,7 @@ package org.nanotek.crawler.service;
 
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.PathValidator;
-import org.nanotek.crawler.data.config.meta.MetaEdge;
+import org.nanotek.crawler.data.domain.meta.MetaEdge;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class SimplePathValidator implements PathValidator<Class<?>, MetaEdge> {
